2022/11/21 2

+++ app/src/main/java/com/example/dku_dow_dpp/Babpickresult2Activity.java
@@ -40,7 +40,7 @@ public class Babpickresult2Activity extends AppCompatActivity {
 
     public void onBackPressed() {
         super.onBackPressed();
-        String myroom_id = current_id;
+        String myroom_id = BabpickmakingActivity.current_id;
         DocumentReference productRef = db.collection("babpick").document("식당별").collection(food).document(myroom_id);
 
         if(productRef != null) {
@@ -64,6 +64,7 @@ public class Babpickresult2Activity extends AppCompatActivity {
         String hour = getstr.getStringExtra("hour");
         String min = getstr.getStringExtra("min");
         current_id = getstr.getStringExtra("myroom_id");
+
         String time = hour + " 시   " + min +" 분";
 
         nametext = findViewById(R.id.nametext);
@@ -71,12 +72,13 @@ public class Babpickresult2Activity extends AppCompatActivity {
         timetext = findViewById(R.id.timetext);
         timetext.setText(time);
 
-        TextView temp = findViewById(R.id.tkkk);
-        temp.setText(eng_name);
 
         int iResId = getResources().getIdentifier( "@drawable/"+eng_name, "drawable", this.getPackageName() );
         place = findViewById(R.id.place);
         place.setImageResource(iResId);
+
+        backbtn = findViewById(R.id.button);
+        backbtn.setOnClickListener(view -> onBackPressed());
 
         returnbtn = findViewById(R.id.return_btn);
         returnbtn.setOnClickListener(new View.OnClickListener() {
@@ -84,19 +86,15 @@ public class Babpickresult2Activity extends AppCompatActivity {
             public void onClick(View view) {
                 Intent intent = new Intent(Babpickresult2Activity.this, MainActivity.class);
 
-                try
-                {
-                    Thread.sleep(1000);
-                } catch (InterruptedException e)
-                {
-                    e.printStackTrace();
-                }
+                String myroom_id = BabpickmakingActivity.current_id;
 
-                String myroom_id = current_id;
-                DocumentReference productRef = db.collection("babpick").document("식당별").collection(food).document(myroom_id);
+                if(myroom_id != null) {
+                    DocumentReference productRef = db.collection("babpick").document("식당별").collection(food).document(myroom_id);
 
-                if(productRef != null) {
-                    productRef.delete();
+                    if (productRef != null) {
+                        productRef.delete();
+                    }
+
                 }
 
                 startActivity(intent);
